requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==1.4.0
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
backoff==2.1.2
beautifulsoup4==4.11.1
branca==0.5.0
CacheControl==0.13.1
cachetools==5.3.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
comtypes==1.1.11
cryptography==37.0.2
dj-database-url==2.1.0
Django==3.2.1
django-ckeditor==6.5.1
django-classy-tags==4.1.0
django-cms==3.11.4
django-cors-headers==4.3.0
django-countries==7.4.2
django-embed-video==1.4.8
django-environ==0.11.2
django-filter==23.3
django-formtools==2.4.1
django-htmlmin==0.11.0
django-js-asset==2.0.0
django-livereload-server==0.4
django-sekizai==4.1.0
django-taggit==3.0.0
django-treebeard==4.7
djangocms-admin-style==3.2.6
djangocms-animate-on-scroll==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docutils==0.18.1
Faker==13.13.0
fcm-django==2.0.0
firebase-admin==6.2.0
folium==0.12.1.post1
frozenlist==1.3.3
google-api-core==2.12.0
google-api-python-client==2.103.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.0
grpcio-status==1.59.0
gTTS==2.2.4
gunicorn==21.2.0
h11==0.12.0
html5lib==1.1
httpcore==0.13.3
httplib2==0.22.0
httpx==0.20.0
idna==3.3
image==1.5.33
Jinja2==3.1.2
keyboard==0.13.5
livereload==2.6.3
MarkupSafe==2.1.1
middleware==1.2.3
msgpack==1.0.7
msoffcrypto-tool==5.0.0
multidict==6.0.4
numpy==1.23.0
olefile==0.46
openai==0.26.5
opencage==2.0.0
opencv-python==4.6.0.66
packaging==23.2
phonenumbers==8.12.51
Pillow==9.1.1
playsound==1.2.2
proto-plus==1.22.3
protobuf==4.24.4
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.12.0
PyJWT==2.8.0
pyOpenSSL==22.0.0
pyparsing==3.1.1
pystache==0.6.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytube==12.1.0
pytz==2022.7
pywifi==1.1.12
PyYAML==6.0
qrcode==7.3.1
requests==2.28.0
rfc3986==1.5.0
rotate-screen==0.1.0
rsa==4.9
scipy==1.8.1
sitemaps==0.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
static==1.1.1
svg.path==6.2
svgpathtools==1.5.1
svgwrite==1.4.3
tornado==6.2
tqdm==4.64.0
typing_extensions==4.4.0
tzdata==2022.1
ua-parser==0.16.1
uritemplate==4.1.1
urllib3==1.26.9
user-agents==2.2.0
webencodings==0.5.1
whitenoise==6.6.0
yarl==1.8.2



# asarPy==1.0.1
# asgiref==3.8.1
# CacheControl==0.14.1
# cachetools==5.5.0
# certifi==2024.8.30
# cffi==1.17.1
# charset-normalizer==3.4.0
# cryptography==44.0.0
# cryptomus==1.1
# Cython==3.0.11
# dj-database-url==2.3.0
# Django==5.1.3
# django-ckeditor==6.7.1
# django-cors-headers==4.6.0
# django-countries==7.6.1
# django-filter==24.3
# django-js-asset==2.2.0
# django-taggit==6.1.0
# djangorestframework==3.15.2
# djangorestframework-simplejwt==5.3.1
# fcm-django==2.2.1
# firebase-admin==6.6.0
# google-api-core==2.24.0
# google-api-python-client==2.155.0
# google-auth==2.37.0
# google-auth-httplib2==0.2.0
# google-cloud-core==2.4.1
# google-cloud-firestore==2.19.0
# google-cloud-storage==2.19.0
# google-crc32c==1.6.0
# google-resumable-media==2.7.2
# googleapis-common-protos==1.66.0
# grpcio==1.68.1
# grpcio-status==1.68.1
# gunicorn==23.0.0
# httplib2==0.22.0
# idna==3.10
# Markdown==3.5.2
# MarkupSafe==2.1.5
# msgpack==1.1.0
# packaging==24.1
# pillow==11.0.0
# proto-plus==1.25.0
# protobuf==5.29.1
# psycopg2-binary==2.9.10
# pyasn1==0.6.1
# pyasn1_modules==0.4.1
# pycairo==1.27.0
# pycparser==2.22
# Pygments==2.18.0
# PyGObject==3.50.0
# PyJWT==2.10.1
# pyparsing==3.2.0
# requests==2.32.3
# rsa==4.9
# setuptools==75.1.0.post20240919
# six==1.16.0
# sqlparse==0.5.2
# swapper==1.4.0
# typing_extensions==4.12.2
# uritemplate==4.1.1
# urllib3==2.2.3
# wheel==0.44.0
# whitenoise==6.8.2